% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_prop_var.R
\name{extract_prop_var}
\alias{extract_prop_var}
\title{Extract the proportion of variance}
\usage{
extract_prop_var(fit)
}
\arguments{
\item{fit}{Output from \code{prcomp}}
}
\value{
Named numeric vector containing the proportion of variance explained
  by each principal component.
}
\description{
This function provides an easy interface for extracing the proportion of
variance from the output of \code{prcomp}.
}
